завершили рефакторинг по первому уроку
